+++ src/test/java/nextstep/courses/domain/ImagePixelTest.java
@@ -9,8 +9,8 @@ public class ImagePixelTest {
     @Test
     @DisplayName("ImageProperty 생성")
     void createImagePropertyTest() {
-        Long width = 300L;
-        Long height = 200L;
+        int width = 300;
+        int height = 200;
 
         ImagePixel imagePixel = new ImagePixel(width, height);
 
@@ -20,8 +20,8 @@ public class ImagePixelTest {
     @Test
     @DisplayName("ImageProperty 최소 width 체크")
     void checkMinWidthTest() {
-        Long width = 299L;
-        Long height = 200L;
+        int width = 299;
+        int height = 200;
 
         Assertions.assertThatThrownBy(() -> new ImagePixel(width, height))
                 .isInstanceOf(IllegalArgumentException.class);
@@ -30,8 +30,8 @@ public class ImagePixelTest {
     @Test
     @DisplayName("ImageProperty 최소 height 체크")
     void checkMinHeightTest() {
-        Long width = 300L;
-        Long height = 199L;
+        int width = 300;
+        int height = 199;
 
         Assertions.assertThatThrownBy(() -> new ImagePixel(width, height))
                 .isInstanceOf(IllegalArgumentException.class);
@@ -40,8 +40,8 @@ public class ImagePixelTest {
     @Test
     @DisplayName("ImageProperty width/height ratio 체크")
     void checkWidthHeightRatioTest() {
-        Long width = 300L;
-        Long height = 300L;
+        int width = 300;
+        int height = 300;
 
         Assertions.assertThatThrownBy(() -> new ImagePixel(width, height))
                 .isInstanceOf(IllegalArgumentException.class);
